exempt variable declarations from checking if in scope
