+++ src/components/sequencing/form/ByteArrayEditor.svelte
@@ -2,8 +2,8 @@
 
 <script lang="ts">
   import type { SyntaxNode } from '@lezer/common';
+  import { decodeInt32Array } from '../../../utilities/codemirror/codemirror-utils';
   import type { CommandInfoMapper } from '../../../utilities/codemirror/commandInfoMapper';
-  import { decodeInt32Array } from '../../../utilities/codemirror/vml/vmlTreeUtils';
 
   export let argNode: SyntaxNode;
   export let commandInfoMapper: CommandInfoMapper;
